+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tokenbound/react
 
+## 0.1.1
+
+### Patch Changes
+
+- Fixed build issues
+
 ## 0.1.0
 
 ### Minor Changes