+++ packages/sdk-ethers/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tokenbound/sdk-ethers
 
+## 0.2.0
+
+### Minor Changes
+
+- Added initialization call to createAccount
+
+## 0.1.1
+
+### Patch Changes
+
+- Fixed build issues
+
 ## 0.1.0
 
 ### Minor Changes