+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @tokenbound/sdk
 
+## 0.2.0
+
+### Minor Changes
+
+- Added initialization call to createAccount
+
+## 0.1.2
+
+### Patch Changes
+
+- Fixed contract address issue
+
+## 0.1.1
+
+### Patch Changes
+
+- Fixed build issues
+
 ## 0.1.0
 
 ### Minor Changes